add logic to delete image when deleting related product

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require("../models/product");
 const { validationResult } = require("express-validator/check");
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
 	res.render("admin/edit-product", {
@@ -148,6 +149,7 @@ exports.postEditProduct = (req, res, next) => {
 			product.price = updatedPrice;
 			product.description = updatedDesc;
 			if (image) {
+				fileHelper.deleteFile(product.imageUrl);
 				product.imageUrl = image.path;
 			}
 			return product.save().then(result => {
@@ -182,7 +184,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
 	const prodId = req.body.productId;
-	Product.deleteOne({ _id: prodId, userId: req.user._id })
+	Product.findById(prodId)
+		.then(prod => {
+			if (!prod) {
+				return next(new Error("Product not found"));
+			}
+			fileHelper.deleteFile(prod.imageUrl);
+			return Product.deleteOne({ _id: prodId, userId: req.user._id });
+		})
 		.then(() => {
 			console.log("DESTROYED PRODUCT");
 			res.redirect("/admin/products");
